Add files via upload
This project is a real-time face recognition system designed with face_recognition, OpenCV, and Flask. Users can utilize it for real-time video streaming via a webcam, face detection, and face recognition with saved photos for authentication. It is web-based, and simple face recognition output tracking can be accomplished through any device in a shared network.